+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- The `release_string_utf_chars` function has been marked as unsafe. (#334)
+
 ## [0.19.0] — 2021-01-24
 
 ### Added